# doc/openapi.yaml
openapi: "3.0.2"
info:
  title: Friend Finder
  version: '1.0'
  description: "L'API de FriendFinder est une API REST. Elle a des URL orientées ressources, elle renvoie des données codées en JSON, elle utilise les codes de réponses standards HTTP, et une authentification."

  #termsOfService: "https://openweathermap.org/terms"
  license:
    name: "Repo Git"
    url: "https://github.com/EnseirbTelecom/pg219-2020-dark-heroes/tree/master/doc"
servers:
  - url: https://localhost:3000
security:
  - friendFinder_auth: []
...
components:
  securitySchemes:
    friendFinder_auth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "L'utilisateur doit s'identifier pour accéder aux requêtes."
  schemas:
    Hist_position:
      type: object
      properties:
        hist_lat:
          type: string
        hist_long:
          type: string
        hist_date:
          type: string
          format: date
        message:
          type: string
    Position:
      type: object
      properties:
        email:
          type: string
          format: email
        pseudo:
          type: string
        lat:
          type: string
        long:
          type: string
        date:
          type: string
          format: date
        duration:
          type: string
          minimum: 0
        message:
          type: string
    User:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        first_name:
          type: string
        last_name:
          type: string
        pseudo:
          type: string
        birthday:
          type: string
          format: date
        gender:
          type: string
        current_lat:
          type: string
        current_long:
          type: string
        activation_date:
          type: string
          format: date
        duration:
          type: number
          minimum: 0
        hist_positions:
          $ref: "#/components/schemas/Hist_position"
        friends:
          type: array
          items:
            type: string
        friend_request:
          type: array
          items:
            type: string
        friend_pending:
          type: array
          items:
            type: string
paths:
...
  /histPosition:
    get:
      tags:
        - Position
      summary: "Renvoie l'historique des positions"
      operationId: "getHistPosition"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  hist_positions:
                    $ref: "#/components/schemas/Hist_position"
                  status:
                    type: integer
                    example: 200
        "403":
          description: "Erreur"
  /currPosition:
    get:
      tags:
        - Position
      summary: "Renvoie la position courante"
      operationId: "getCurrentPosition"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  position:
                    $ref: "#/components/schemas/Position"
                  status:
                    type: integer
                    example: 200
        "201":
          description: "Utilisateur introuvable"
          content:
            json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
        "403":
          description: "Erreur"
  /currFriendsPosition:
    get:
      tags:
        - Position
      summary: "Renvoie la position courrente d'un ami"
      operationId: "getFriendsCurrentPosition"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  friendsposition:
                    $ref: "#/components/schemas/Position"
                  status:
                    type: integer
                    example: 200
        "201":
          description: "Position indisponible, ou ami introuvable"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "no friends position are available"
                  status:
                    type: integer
                    example: 201
  /savePos:
    post:
      tags:
        - Position
      summary: "Sauvegarder une position"
      operationId: "addPosition"
      requestBody:
        description: "Structure User de l'utilisateur"
        required: true
        content:
          json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "Position added"
                  status:
                    type : integer
                    example: 200
        "403":
          description: "Erreur"
  /delPos:
    delete:
      tags:
        - Position
      summary: "Supprimer une position"
      operationId: "deletePosition"
      requestBody:
        description: "Structure User de l'utilisateur"
        required: true
        content:
          json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "element deleted"
                  status:
                    type: integer
                    example: 200
        "403":
          description: "Erreur"
...